Move 'reduce' method to Match class
